Renamed files to use a folder base test stucture and moved them accordingly.
